split-identified-facet-type
