+++ toolchain/sem_ir/BUILD
@@ -98,6 +98,7 @@ cc_library(
         "file.cpp",
         "function.cpp",
         "generic.cpp",
+        "identified_facet_type.cpp",
         "impl.cpp",
         "import_ir.cpp",
         "inst.cpp",
@@ -132,6 +133,7 @@ cc_library(
         "file.h",
         "function.h",
         "generic.h",
+        "identified_facet_type.h",
         "impl.h",
         "import_cpp.h",
         "import_ir.h",
